Module parser::combinators

Provided combinators

Functions

count
Matches the given parser exactly num times, returning all the matches.

many
Parses many instances of f until it does no longer match, returning all matches.

many1
Parses at least one instance of f and continues until it does no longer match, returning all matches.

option
Tries the parser f, on success it yields the parsed value, on failure default will be yielded instead.

or
Tries to match the parser f, if f fails it tries g. Returns the success value of the first match, otherwise the error of the last one if both fail.

skip_many
Runs the given parser until it fails, discarding matched input.