Function parser::combinators::many

pub fn many<'a, I: 'a + Copy, T, E, F, U>(m: Empty<'a, I>, f: F) -> Parser<'a, I, T, E> where F: Fn(Empty<'a, I>) -> Parser<'a, I, U, E>, T: FromIterator<U>

Parses many instances of f until it does no longer match, returning all matches.

Note: If the last parser succeeds on the last input item then this parser is still considered incomplete as there might be more data to fill.

Note: Allocates data.

use parser::{Error, Parser, bind, char, many, ret, take_while1};
 
let p = From::from(b"a,bc,cd ");
 
let r: Parser<_, Vec<&[u8]>, Error<_>> = many(p, |m| bind(
    take_while1(m, |c| c != b',' && c != b' '), |m, c| bind(
        char(m, b','), |m, _| ret(m, c))));
let v = r.unwrap();

assert_eq!(v.len(), 2);
assert_eq!(v[0], b"a");
assert_eq!(v[1], b"bc");