Function parser::combinators::skip_many

pub fn skip_many<'a, I: 'a + Copy, T, E, F>(m: Empty<'a, I>, f: F) -> Parser<'a, I, (), Error<I>> where F: Fn(Empty<'a, I>) -> Parser<'a, I, T, E>

Runs the given parser until it fails, discarding matched input.

Incomplete state will be propagated.

This is more efficient to use compared to using many and then just discarding the result as many allocates a separate data structure to contain the data before proceeding.

use parser::{Parser, bind, skip_many, char};
 
let p = From::from(b"aaaabc");
 
assert_eq!(bind(skip_many(p, |m| char(m, b'a')), |m, _| char(m, b'b')).unwrap(), b'b');