Enum parser::OpArity

pub enum OpArity {
    Binary,
    Literal,
    Unary,
    Function,
}

OpArity is an enumeration that represents the four ways identifiers and operators can be used.

Binary: Two operands, on the left and right. This is only macros for Alnozo. Literal: A literal, atomic value. Unary: An operator that has only one operand, on the right. Funciton: Either a macro or a function called normally. TODO: Rename this.

Variants

Binary

Literal

Unary

Function