Function parser::functor::map

pub fn map<'a, I: 'a + Copy, T, E, F, U>(m: Parser<'a, I, T, E>, f: F) -> Parser<'a, I, U, E> where F: FnOnce(T) -> U

Applies the function f to the wrapped value if the parser is in a success state, otherwise does nothing.

use parser::{Error, Parser, map, ret};
 
let p: Parser<_, usize, Error<()>> = ret(From::from(b"dummy"), 1234);
 
assert_eq!(map(p, |i| i + 20).unwrap(), 1254);