Function parser::monad::err

pub fn err<'a, I: 'a + Copy, T, E>(m: Empty<'a, I>, err: E) -> Parser<'a, I, T, E>

Constructs an error value from the given error err.

Currently this function may need type-annotation for the second type-parameter T because the success-value is unconstrained by this function. When wrapped in smaller functions inference should be able to use the return-value of the function to infer return and error types.

use parser::{Parser, Error, err};
 
let p = From::from(b"test");
 
assert_eq!(err::<_, &[u8], _>(p, Error::Unexpected(2)).unwrap(), b"data");