parser::parser

Struct Parser

pub(crate) struct Parser<'t> {
    inp: &'t Input,
    pos: usize,
    events: Vec<Event>,
    steps: Cell<u32>,
}

Parser struct provides the low-level API for navigating through the stream of tokens and constructing the parse tree. The actual parsing happens in the grammar module.

However, the result of this Parser is not a real tree, but rather a flat stream of events of the form “start expression, consume number literal, finish expression”. See Event docs for more.

Fields
inp: &'t Inputpos: usizeevents: Vec<Event>steps: Cell<u32>
Implementations

impl<'t> Parser<'t>

pub(crate) fn new(inp: &'t Input) -> Parser<'t>

pub(crate) fn finish(self) -> Vec<Event>

pub(crate) fn current(&self) -> SyntaxKind

Returns the kind of the current token. If parser has already reached the end of input, the special EOF kind is returned.

pub(crate) fn nth(&self, n: usize) -> SyntaxKind

Lookahead operation: returns the kind of the next nth token.

pub(crate) fn at(&self, kind: SyntaxKind) -> bool

Checks if the current token is kind.

pub(crate) fn nth_at(&self, n: usize, kind: SyntaxKind) -> bool

pub(crate) fn eat(&mut self, kind: SyntaxKind) -> bool

Consume the next token if kind matches.

pub(crate) fn eat_contextual_kw(&mut self, kind: SyntaxKind) -> bool

fn at_composite2(&self, n: usize, k1: SyntaxKind, k2: SyntaxKind) -> bool

fn at_composite3( &self, n: usize, k1: SyntaxKind, k2: SyntaxKind, k3: SyntaxKind, ) -> bool

pub(crate) fn at_ts(&self, kinds: TokenSet) -> bool

Checks if the current token is in kinds.

pub(crate) fn at_contextual_kw(&self, kw: SyntaxKind) -> bool

Checks if the current token is contextual keyword kw.

pub(crate) fn nth_at_contextual_kw(&self, n: usize, kw: SyntaxKind) -> bool

Checks if the nth token is contextual keyword kw.

pub(crate) fn start(&mut self) -> Marker

Starts a new node in the syntax tree. All nodes and tokens consumed between the start and the corresponding Marker::complete belong to the same node.

pub(crate) fn bump(&mut self, kind: SyntaxKind)

Consume the next token. Panics if the parser isn’t currently at kind.

pub(crate) fn bump_any(&mut self)

Advances the parser by one token

pub(crate) fn split_float(&mut self, marker: Marker) -> (bool, Marker)

Advances the parser by one token

pub(crate) fn bump_remap(&mut self, kind: SyntaxKind)

Advances the parser by one token, remapping its kind. This is useful to create contextual keywords from identifiers. For example, the lexer creates a union identifier token, but the parser remaps it to the union keyword, and keyword is what ends up in the final tree.

pub(crate) fn error<T: Into<String>>(&mut self, message: T)

Emit error with the message FIXME: this should be much more fancy and support structured errors with spans and notes, like rustc does.

pub(crate) fn expect(&mut self, kind: SyntaxKind) -> bool

Consume the next token if it is kind or emit an error otherwise.

pub(crate) fn err_and_bump(&mut self, message: &str)

Create an error node and consume the next token.

pub(crate) fn err_recover(&mut self, message: &str, recovery: TokenSet) -> bool

Create an error node and consume the next token unless it is in the recovery set.

Returns true if recovery kicked in.

fn do_bump(&mut self, kind: SyntaxKind, n_raw_tokens: u8)

fn push_event(&mut self, event: Event)

pub(crate) fn current_edition(&self) -> Edition