Module parser::parsers

Provided parsers.

Functions

any
Matches any item, returning it if present.

char
Matches a single character, returning the matched character on success.

not_char
Matches an item as long as it is not equal to c.

peek
Matches any item but does not consume it, on success it gives Some but if no input remains None is produced.

satisfy
Matches an item using f, the item is returned if f yields true, otherwise this parser fails.

string
Matches the given slice against the parser, returning the matched slice upon success.

take
Matches num items no matter what they are, returning a slice of the matched items.

take_remainder
Matches the remainder of the buffer and returns it, always succeeds.

take_till
Matches all items until f returns true, all items to that point will be returned as a slice upon success.

take_while
Matches all items while f returns false, returns a slice of all the matched items.

take_while1
Matches all items while f returns true, if at least one item matched this parser succeeds and returns a slice of all the matched items.