Function parser::parsers::peek

pub fn peek<'a, I: 'a + Copy>(m: Empty<'a, I>) -> Parser<'a, I, Option<I>, Error<I>>

Matches any item but does not consume it, on success it gives Some but if no input remains None is produced.

This parser is never considered incomplete.

use parser::{Parser, peek};
 
let p1 = From::from(b"abc");
let p2 = From::from(b"");
 
assert_eq!(peek(p1).unwrap(), Some(b'a'));
assert_eq!(peek(p2).unwrap(), None);