Function parser::parsers::satisfy

pub fn satisfy<'a, I: 'a + Copy, F>(m: Empty<'a, I>, f: F) -> Parser<'a, I, I, Error<I>> where F: FnOnce(I) -> bool

Matches an item using f, the item is returned if f yields true, otherwise this parser fails.

If the buffer length is 0 this parser is considered incomplete.

use parser::{Parser, satisfy};
 
let p = From::from(b"abc");
 
assert_eq!(satisfy(p, |c| c == b'a').unwrap(), b'a');