parser
Struct Output

pub struct Output { /* private fields */ }

Output of the parser – a DFS traversal of a concrete syntax tree.

Use the Output::iter method to iterate over traversal steps and consume a syntax tree.

In a sense, this is just a sequence of SyntaxKind-colored parenthesis interspersed into the original crate::Input. The output is fundamentally coordinated with the input and n_input_tokens refers to the number of times crate::Input::push was called.

Implementations

impl Output

pub fn iter(&self) -> impl Iterator<Item = Step<'_>>