Struct parser::Parser

#[must_use]
pub struct Parser<'a, I: 'a + Copy, T, E>(_, _);

The main parser data-type, contains the current fragment to be parsed and the current parser-state.

Methods

impl<'a, I: 'a + Copy, T, E> Parser<'a, I, T, E>

fn map<F, U>(self, f: F) -> Parser<'a, I, U, E> where F: FnOnce(T) -> U

Applies the function f to the wrapped value if the parser is in a success state, otherwise does nothing.

use parser::{Error, Parser, ret};
 
let p: Parser<_, usize, Error<()>> = ret(From::from(b"dummy"), 1234);
 
assert_eq!(p.map(|i| i + 20).unwrap(), 1254);

impl<'a, I: 'a + Copy + Debug, T, E: Error> Parser<'a, I, T, E>

fn unwrap(self) -> T

Extracts the contained type if the parser is in a success-state, panics otherwise.