Trait parser::TreeSink

pub trait TreeSink {
    type FinishResult;
    fn token(&mut self);

    fn start_node(&mut self, kind: SyntaxKind);

    fn finish_node(&mut self);

    fn start_error_node(&mut self);

    fn finish_error_node(&mut self, e: ParseError);

    fn error(&mut self, error: ParseError);

    fn finish(self) -> Self::FinishResult;
}

TreeSink abstracts details of a particular syntax tree implementation.

Associated Types

type FinishResult

Loading content...
Required methods

fn token(&mut self)

Adds new token to the current branch.

fn start_node(&mut self, kind: SyntaxKind)

Start new branch and make it current.

fn finish_node(&mut self)

Finish current branch and restore previous branch as current.

fn start_error_node(&mut self)

fn finish_error_node(&mut self, e: ParseError)

fn error(&mut self, error: ParseError)

fn finish(self) -> Self::FinishResult

Loading content...