Crate parser

parser

findrepl

The provided code updates a section of a file by replacing the content between specified start and end markers with a given text. The code opens the file, reads its contents into a buffer, finds the start and end markers, removes the content between them, and inserts the updated text. The updated content is then written back to the file and the function returns a Result.

Re-exports

pub use crate::findrepl::*;

Modules

findrepl

findrepl

Macros

comment_block

The macro comment_block generates the start and end marker strings of a comment section in a Markdown file.

comment_block_dyn

Macro comment_block_dyn macro accepts three arguments: $section_name, $marker, and $section. The $section argument is used to dynamically specify the section name in the generated string. The $section_name and $marker arguments work just like in the previous version.

Enums

ErrorColor

The ErrorColor enum represents the different colors that can be used.

ParserError

The ParserError enum represents the different errors that can occur while parsing some input.

PrinterError

PrinterError enum represents the different errors that can occur while printing some output.

Type Definitions

Result

Result<T, E> is an alias for anyhow::Result with ParserError as the error type. It is used as the return type for functions that may fail and return an error.