Crate parser

... is an eager zero-copy monadic parser combinator operating on arbitrary slices of data.

Reexports

pub use functor::map;

pub use monad::{bind, err, ret};

pub use combinators::{option, count, or, many, many1, skip_many};

pub use parsers::{char, not_char, satisfy, peek, take, take_till, take_while, take_while1, take_remainder, string};

Modules

any
Traits for dynamic typing of any 'static type (through runtime reflection)

combinators
Provided combinators

functor
Functor implementation for Parser.

iter

monad
Monad implementation for Parser.

parsers
Provided parsers.

Macros

mdo!
Macro for using monadic do like haskell's do-notation.

Structs

Parser
The main parser data-type, contains the current fragment to be parsed and the current parser-state.

Enums

Error

Type Definitions

Empty
A parser state where no data or error is present.