Crate parser

Macros

N
A helper macro to get the node

T
A helper macro to get the token

TS

Structs

Reparser
A parsing function for a block. Blocks can be from functions, do stmts, loops, etc. Anything that has an end.

Token
A lua token

TokenSet
A bit-set of SyntaxKinds

Enums

Event
TODO: add node error event

LuaOp

ParseError

SyntaxKind
The kind of syntax node, e.g. Ident, UseKw, or Struct.

Traits

TokenSource
Token abstracts the cursor of TokenSource operates on.

TreeSink
TreeSink abstracts details of a particular syntax tree implementation.

Functions

parse
Parse given tokens into the given sink as a rust file.