Module parser::findrepl

findrepl

The code defines a CommentBlock struct to store a section name and its start and end markers, which can be generated from the section name. A comment_block_marker macro is defined to generate a marker string for a section with a specified marker type.

An example of how to use the function try_lib_main is provided, which demonstrates how it can be used to replace the content between start and end markers in a file with a new text.

Example

Input - example.txt

# findreplace

## List

<!--START_SECTION:tag_1-->
* [lorem](https://github.com/username/username) — username's GitHub profile.
* [foobar](https://github.com/username/foobar)
* [bar](https://github.com/username/bar) — Lorem ipsum dolor sit amet, qui minim labore
  adipisicing minim sint cillum sint consectetur cupidatat. ...
<!--END_SECTION:tag_1-->

fn main() {
    let path = Path::new("example.txt");
    let text = "* [new_lorem](https://github.com/new_username/new_username) — new_username's GitHub profile.\n\
                * [new_foobar](https://github.com/new_username/new_foobar)\n\
                * [new_bar](https://github.com/new_username/new_bar) — Lorem ipsum dolor sit amet, qui minim labore\n\
                  adipisicing minim sint cillum sint consectetur cupidatat.";
    if let Err(error) = try_lib_main(text, &path) {
        println!("Error: {}", error);
    }
}

The function try_lib_main will then update the file by replacing the content between the start and end markers and with the string text. #

Output:

# findreplace

## List

<!--START_SECTION:tag_1-->
* [new_lorem](https://github.com/new_username/new_username) — new_username's GitHub
  profile.\n\
* [new_foobar](https://github.com/new_username/new_foobar)\n\
* [new_bar](https://github.com/new_username/new_bar) — Lorem ipsum dolor sit amet, qui minim
  labore\n\ adipisicing minim sint cillum sint consectetur cupidatat."; ...
<!--END_SECTION:tag_1-->

Structs

CommentBlock

CommentBlock is a struct that holds information about a comment block. It has two fields: section_name, which is a String representing the name of the section, and marker, which is a tuple of two Marker values, indicating the start and end markers of the comment block.

Enums

Marker

Marker is an enumeration of marker values, Start and End. These markers are used to indicate the start and end of a comment block in some implementation.

Functions

replace

try_main